requirements.txt
wheel
pybluez
beacontools
pyserial
